showing on user page all the videos of the user

+++ server/services/videoService.js
@@ -43,9 +43,19 @@ const updateVideo = async (userId, videoId, updateData) => {
   return await Video.findOneAndUpdate({ _id: videoId, 'uploader.id': userId }, updateData, { new: true });
 };
 
+const mongoose = require('mongoose');
+
 const deleteVideo = async (userId, videoId) => {
-  return await Video.findOneAndDelete({ _id: videoId, 'uploader.id': userId });
+  console.log(`Attempting to delete video with ID: ${videoId} for user ID: ${userId}`);
+  if (!mongoose.Types.ObjectId.isValid(videoId) || !mongoose.Types.ObjectId.isValid(userId)) {
+    throw new Error('Invalid user ID or video ID');
+  }
+  
+  const result = await Video.findOneAndDelete({ _id: videoId, 'uploader.id': userId });
+  console.log('Delete result:', result);
+  return result;
 };
+
 
 module.exports = {
   getVideos,
